@@ -1,83 +1,66 @@
 {
   "name": "colorpeekr",
-  "version": "0.2.0",
-  "private": true,
-  "main": "build/browser.js",
+  "productName": "colorpeekr",
+  "version": "0.3.0",
+  "main": ".webpack/main",
   "scripts": {
-    "dev": "electron-forge start",
+    "start": "electron-forge start",
+    "package": "electron-forge package",
     "make": "electron-forge make",
-    "build": "webpack",
-    "watch": "node ./scripts/watchFs.js",
-    "package": "electron-forge package"
+    "publish": "electron-forge publish",
+    "rebuild": "npx electron-rebuild -v 8.2.4 -f",
+    "lint": "echo \"No linting configured\""
   },
-  "dependencies": {
-    "classnames": "^2.2.6",
-    "color-namer": "^1.3.0",
-    "electron-compile": "^6.4.2",
-    "lodash": "^4.17.4",
-    "mobx": "^5.0.3",
-    "mobx-react": "^5.2.3",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "robotjs": "^0.4.7"
+  "keywords": [],
+  "author": {
+    "name": "theadriann",
+    "email": "adrian.brojbeanu@gmail.com"
+  },
+  "license": "MIT",
+  "config": {
+    "forge": "./configs/forge.config.json"
   },
   "devDependencies": {
-    "autoprefixer": "^7.2.5",
-    "babel-eslint": "^8.0.1",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.5",
-    "babel-plugin-transform-function-bind": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "clean-webpack-plugin": "^0.1.17",
-    "css-loader": "^0.28.9",
-    "electron-forge": "^4.1.2",
-    "electron-prebuilt-compile": "1.8.2-beta.3",
-    "electron-squirrel-startup": "^1.0.0",
-    "eslint": "^4.16.0",
-    "eslint-plugin-babel": "^4.1.2",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.6",
-    "html-webpack-plugin": "^2.30.1",
-    "less": "^3.0.0-alpha.3",
-    "less-loader": "^4.0.5",
+    "@babel/core": "^7.9.6",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "^7.8.3",
+    "@babel/plugin-proposal-function-sent": "^7.8.3",
+    "@babel/plugin-proposal-numeric-separator": "^7.8.3",
+    "@babel/plugin-proposal-throw-expressions": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
+    "@electron-forge/cli": "^6.0.0-beta.51",
+    "@electron-forge/maker-deb": "^6.0.0-beta.51",
+    "@electron-forge/maker-rpm": "^6.0.0-beta.51",
+    "@electron-forge/maker-squirrel": "^6.0.0-beta.51",
+    "@electron-forge/maker-zip": "^6.0.0-beta.51",
+    "@electron-forge/plugin-auto-unpack-natives": "^6.0.0-beta.51",
+    "@electron-forge/plugin-webpack": "^6.0.0-beta.51",
+    "@marshallofsound/webpack-asset-relocator-loader": "^0.5.0",
+    "autoprefixer": "^9.7.6",
+    "babel-loader": "^8.1.0",
+    "classnames": "^2.2.6",
+    "color-namer": "^1.4.0",
+    "css-loader": "^3.5.3",
+    "electron": "8.2.4",
+    "electron-rebuild": "^1.10.1",
+    "file-loader": "^6.0.0",
+    "less": "^3.11.1",
+    "less-loader": "^6.0.0",
+    "lodash": "^4.17.15",
+    "mobx": "^5.15.4",
+    "mobx-react": "^6.2.2",
     "node-loader": "^0.6.0",
-    "postcss-loader": "^2.0.10",
-    "strip-ansi": "^4.0.0",
-    "style-loader": "^0.19.1",
-    "webpack": "^3.7.1",
-    "webpack-dev-server": "^2.11.1"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-hot-loader": "^4.12.21",
+    "style-loader": "^1.2.1"
   },
-  "config": {
-    "forge": {
-      "make_targets": {
-        "win32": [
-          "zip"
-        ],
-        "darwin": [
-          "zip"
-        ]
-      },
-      "electronPackagerConfig": {
-        "name": "ColorPeekr",
-        "icon": "./resources/colorpeekr"
-      },
-      "electronWinstallerConfig": {
-        "name": "ColorPeekr",
-        "authors": "theadriann",
-        "description": "An open-source color picker built with web technology"
-      },
-      "electronInstallerDebian": {},
-      "electronInstallerRedhat": {},
-      "github_repository": {
-        "owner": "",
-        "name": ""
-      },
-      "windowsStoreConfig": {
-        "packageName": ""
-      }
-    }
+  "dependencies": {
+    "electron-squirrel-startup": "^1.0.0",
+    "robotjs": "^0.6.0"
   }
 }
